Determine when to stop by change in bounding box
Since all the points move linearly, the bounding box will shrink and
then grow without bound in a roughly hyperbolic fashion (although the x
and y waists could be in two different places, and the bounding box
could also stay the same but if that's the case then there is no
relative motion of the points, so the initial positions will give the
message).  This phenomenon is used as a heuristic to determine when to
stop advancing the points.  It assumes that the x and y waists occur
simultaneously and that the message is also readable at this point.